refactor(frontend/src/pages/PredictionEvents/partials/OccultationMap/FlyToMap.js-frontend/src/pages/PredictionEvents/partials/OccultationMap/Legend.js-frontend/src/pages/PredictionEvents/partials/OccultationMap/NightTime.js-frontend/src/pages/PredictionEvents/partials/OccultationMap/index.js-frontend/src/components/OccultationMap/): Os componentes utiizados no mapa foram movidos da pasta PredictionEvents/Occultation para uma pasta especifica em componentes (componentes/Occultation)
