+++ frontend/src/components/OccultationMap/Legend.js
