+++ frontend/src/components/OccultationMap/NightTime.js
